+++ pyproject.toml
@@ -80,6 +80,9 @@ skip-magic-trailing-comma = false
 line-ending = "auto"
 docstring-code-format = false
 
+[tool.ruff.lint.per-file-ignores]
+"sudoku/__init__.py" = ["F401"]
+
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
